Allow Judges to see all information, regardless of room settings

+++ common/server_cardzone.cpp
@@ -323,7 +323,7 @@ void Server_CardZone::getInfo(ServerInfo_Zone *info, Server_Player *playerWhosAs
     info->set_name(name.toStdString());
     info->set_type(type);
     info->set_with_coords(has_coords);
-    info->set_card_count(cards.size());
+    info->set_card_count(static_cast<int>(cards.size()));
     info->set_always_reveal_top_card(alwaysRevealTopCard);
     info->set_always_look_at_top_card(alwaysLookAtTopCard);
 
@@ -331,7 +331,6 @@ void Server_CardZone::getInfo(ServerInfo_Zone *info, Server_Player *playerWhosAs
     const auto zonesSelfCanSee = type != ServerInfo_Zone::HiddenZone;
     const auto otherPlayerAsking = playerWhosAsking != player;
     const auto zonesOthersCanSee = type == ServerInfo_Zone::PublicZone;
-
     if ((selfPlayerAsking && zonesSelfCanSee) || (otherPlayerAsking && zonesOthersCanSee)) {
         QListIterator<Server_Card *> cardIterator(cards);
         while (cardIterator.hasNext())
